fix: fix display images in catalog page and change save refresh token
