+++ src/views/user-profile/Api-Userprofile.tsx
@@ -1,35 +1,40 @@
 import { errorNotify } from '../../utils/ErrorPupUp';
 import { successNotify } from '../../utils/SuccessPopUp';
+import { scheduleTokenRefresh } from '../../utils/refreshToken';
 
 export async function setNewFirstName(version: string, firstName: string) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'setFirstName',
-              firstName: firstName,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      successNotify(`Your name has been successfully changed to ${firstName}`);
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`New name set successfully`);
-    } else {
-      console.error(`Failed to set new name`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'setFirstName',
+                firstName: firstName,
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        successNotify(
+          `Your name has been successfully changed to ${firstName}`
+        );
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error setting new name:`, error);
@@ -38,35 +43,37 @@ export async function setNewFirstName(version: string, firstName: string) {
 
 export async function setNewLastName(version: string, lastName: string) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'setLastName',
-              lastName: lastName,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      successNotify(
-        `Your last name has been successfully changed to ${lastName}`
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'setLastName',
+                lastName: lastName,
+              },
+            ],
+          }),
+        }
       );
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`New last name set successfully`);
-    } else {
-      console.error(`Failed to set new last name`);
+
+      if (response.ok) {
+        const data = await response.json();
+        successNotify(
+          `Your last name has been successfully changed to ${lastName}`
+        );
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error setting new last name:`, error);
@@ -75,35 +82,37 @@ export async function setNewLastName(version: string, lastName: string) {
 
 export async function setNewDateOfBirth(version: string, dateOfBirth: string) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'setDateOfBirth',
-              dateOfBirth: dateOfBirth,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      successNotify(
-        `Your date of birth has been successfully changed to ${dateOfBirth}`
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'setDateOfBirth',
+                dateOfBirth: dateOfBirth,
+              },
+            ],
+          }),
+        }
       );
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`New date of birth set successfully`);
-    } else {
-      console.error(`Failed to set new date of birth`);
+
+      if (response.ok) {
+        const data = await response.json();
+        successNotify(
+          `Your date of birth has been successfully changed to ${dateOfBirth}`
+        );
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error setting new date of birth:`, error);
@@ -112,33 +121,35 @@ export async function setNewDateOfBirth(version: string, dateOfBirth: string) {
 
 export async function setNewEmail(version: string, email: string) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'changeEmail',
-              email: email,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      successNotify(`Your email has been successfully changed to ${email}`);
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`New email set successfully`);
-    } else {
-      console.error(`Failed to set new email`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'changeEmail',
+                email: email,
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        successNotify(`Your email has been successfully changed to ${email}`);
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error setting new email:`, error);
@@ -151,32 +162,35 @@ export async function resetPassword(
   currentPassword: string
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me/password`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          currentPassword: currentPassword,
-          newPassword: newPassword,
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    const data = await response.json();
-    if (response.ok) {
-      successNotify(`Your passwor has been successfully changed`);
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`New password set successfully`);
-    } else {
-      if (data.statusCode === 400) {
-        errorNotify(`Wrong current password`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me/password`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            currentPassword: currentPassword,
+            newPassword: newPassword,
+          }),
+        }
+      );
+
+      const data = await response.json();
+      if (response.ok) {
+        successNotify(`Your passwor has been successfully changed`);
+        localStorage.setItem('customer', JSON.stringify(data));
+      } else {
+        if (data.statusCode === 400) {
+          errorNotify(`Wrong current password`);
+        }
       }
-      console.error(`Failed to set new password`);
     }
   } catch (error) {
     console.error(`Error setting new password:`, error);
@@ -192,39 +206,41 @@ export async function changeAddress(
   country: string | null
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'changeAddress',
-              addressId: addressId,
-              address: {
-                streetName: street,
-                postalCode: postalCode,
-                city: city,
-                country: country?.slice(country.indexOf('(') + 1, -1),
-              },
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      successNotify(`Your address has been successfully changed`);
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`Address change successfully`);
-    } else {
-      console.error(`Failed to change address`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'changeAddress',
+                addressId: addressId,
+                address: {
+                  streetName: street,
+                  postalCode: postalCode,
+                  city: city,
+                  country: country?.slice(country.indexOf('(') + 1, -1),
+                },
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        successNotify(`Your address has been successfully changed`);
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error changing address:`, error);
@@ -237,33 +253,34 @@ async function setShippingAddress(
   type: string
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: `addShippingAddressId`,
-              addressId: `${addressId}`,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      console.log(data);
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`${type} seted successfully`);
-    } else {
-      console.error(`Failed to set ${type}`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: `addShippingAddressId`,
+                addressId: `${addressId}`,
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error setting ${type}:`, error);
@@ -276,32 +293,34 @@ async function setBillingAddress(
   type: string
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: `addBillingAddressId`,
-              addressId: `${addressId}`,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`${type} seted successfully`);
-    } else {
-      console.error(`Failed to set ${type}`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: `addBillingAddressId`,
+                addressId: `${addressId}`,
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        localStorage.setItem('customer', JSON.stringify(data));
+      }
     }
   } catch (error) {
     console.error(`Error setting ${type}:`, error);
@@ -317,51 +336,53 @@ export async function addAddresses(
   type: string
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'addAddress',
-              address: {
-                streetName: `${streetName}`,
-                postalCode: `${postalCode}`,
-                city: `${city}`,
-                country: country?.slice(country.indexOf('(') + 1, -1),
-              },
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      console.log(`New addresses added successfully`);
-      successNotify(`Your new address has been successfully added`);
-      if (type === 'billing') {
-        await setBillingAddress(
-          data.addresses[data.addresses.length - 1].id,
-          data.version,
-          type
-        );
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'addAddress',
+                address: {
+                  streetName: `${streetName}`,
+                  postalCode: `${postalCode}`,
+                  city: `${city}`,
+                  country: country?.slice(country.indexOf('(') + 1, -1),
+                },
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        successNotify(`Your new address has been successfully added`);
+        if (type === 'billing') {
+          await setBillingAddress(
+            data.addresses[data.addresses.length - 1].id,
+            data.version,
+            type
+          );
+        }
+        if (type === 'shipping') {
+          await setShippingAddress(
+            data.addresses[data.addresses.length - 1].id,
+            data.version,
+            type
+          );
+        }
       }
-      if (type === 'shipping') {
-        await setShippingAddress(
-          data.addresses[data.addresses.length - 1].id,
-          data.version,
-          type
-        );
-      }
-    } else {
-      console.error(`Failed to add new addresses`);
     }
   } catch (error) {
     console.error(`Error adding new addresses:`, error);
@@ -370,33 +391,35 @@ export async function addAddresses(
 
 export async function removeAddresses(version: string, addressId: string) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: 'removeAddress',
-              addressId: addressId,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`Addresses remove successfully`);
-      successNotify(`Your address has been successfully removed`);
-    } else {
-      console.error(`Failed to remove addresses`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: 'removeAddress',
+                addressId: addressId,
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        localStorage.setItem('customer', JSON.stringify(data));
+        successNotify(`Your address has been successfully removed`);
+      }
     }
   } catch (error) {
     console.error(`Error remove addresses:`, error);
@@ -408,35 +431,37 @@ export async function setDefaultShippingAddress(
   addressId: string
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: `setDefaultShippingAddress`,
-              addressId: `${addressId}`,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      console.log(`Shipping default address seted successfully`);
-      localStorage.setItem('customer', JSON.stringify(data));
-      successNotify(
-        `Your default shipping address has been successfully seted`
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: `setDefaultShippingAddress`,
+                addressId: `${addressId}`,
+              },
+            ],
+          }),
+        }
       );
-    } else {
-      console.error(`Failed to set shipping default address`);
+
+      if (response.ok) {
+        const data = await response.json();
+        localStorage.setItem('customer', JSON.stringify(data));
+        successNotify(
+          `Your default shipping address has been successfully seted`
+        );
+      }
     }
   } catch (error) {
     console.error(`Error setting shipping default address:`, error);
@@ -448,33 +473,37 @@ export async function setDefaultBillingAddress(
   addressId: string
 ) {
   try {
-    const response = await fetch(
-      `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${localStorage.getItem('authToken')}`,
-        },
-        body: JSON.stringify({
-          version: version,
-          actions: [
-            {
-              action: `setDefaultBillingAddress`,
-              addressId: `${addressId}`,
-            },
-          ],
-        }),
-      }
-    );
+    await scheduleTokenRefresh();
+    const authToken = localStorage.getItem('authToken');
 
-    if (response.ok) {
-      const data = await response.json();
-      localStorage.setItem('customer', JSON.stringify(data));
-      console.log(`Billing default address seted successfully`);
-      successNotify(`Your default billing address has been successfully seted`);
-    } else {
-      console.error(`Failed to set billing default address`);
+    if (authToken) {
+      const response = await fetch(
+        `${process.env.REACT_APP_CTP_API_URL}/${process.env.REACT_APP_CTP_PROJECT_KEY}/me`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${authToken}`,
+          },
+          body: JSON.stringify({
+            version: version,
+            actions: [
+              {
+                action: `setDefaultBillingAddress`,
+                addressId: `${addressId}`,
+              },
+            ],
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        localStorage.setItem('customer', JSON.stringify(data));
+        successNotify(
+          `Your default billing address has been successfully seted`
+        );
+      }
     }
   } catch (error) {
     console.error(`Error setting billing default address:`, error);
